+++ RELEASE_NOTES.md
@@ -1,2 +1,3 @@
 v0.99.2
 
+- node: Expose Svr2BackupSession serialize()/deserialize()